requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# HTTP provider for API calls
apache-airflow-providers-http==4.5.0

# PostgreSQL provider for database operations
apache-airflow-providers-postgres==5.7.0

# BigQuery provider for Google Cloud BigQuery operations - updated to compatible version
apache-airflow-providers-google==10.8.0

# AWS provider for S3/MinIO operations - removed due to compatibility issues
# apache-airflow-providers-amazon==8.10.0

# boto3 for direct S3/MinIO operations
boto3==1.34.0

# psycopg2-binary==2.9.7